Adds ktlint, configures it, and makes fixes per ktlint check (#9)
* Adds ktlint, configures it, and makes fixes per ktlint check

* Migrate all kotlin only source from Java to Kotlin directories

* ktlint fixes
