+++ coroutines/src/test/kotlin/com/frybits/rx/preferences/coroutines/CoroutinePreferenceTest.kt
