+++ libs.versions.toml
@@ -41,6 +41,7 @@ mockitoKotlin = "4.1.0"
 
 publish-plugin = "1.1.0"
 dependency-analysis = "1.19.0"
+ktlint-plugin = "11.3.1"
 
 [libraries]
 ## Build libraries
@@ -50,6 +51,7 @@ kotlin-gradle-plugin = { module = "org.jetbrains.kotlin:kotlin-gradle-plugin", v
 kotlin-dokka-plugin = { module = "org.jetbrains.dokka:dokka-gradle-plugin", version.ref = "dokka" }
 androidx-navigation-safeargs-plugin = { module = "androidx.navigation:navigation-safe-args-gradle-plugin", version.ref = "androidx-navigation" }
 hilt-plugin = { module = "com.google.dagger:hilt-android-gradle-plugin", version.ref = "hilt" }
+ktlint-plugin = { module = "org.jlleitschuh.gradle:ktlint-gradle", version.ref = "ktlint-plugin" }
 
 ## Project libraries
