+++ livedata/src/test/kotlin/com/frybits/rx/preferences/livedata/PointPreferenceConverter.kt
@@ -1,4 +1,6 @@
-package com.frybits.rx.preferences.core
+package com.frybits.rx.preferences.livedata
+
+import com.frybits.rx.preferences.core.Preference
 
 /*
  *  Copyright 2014 Prateek Srivastava
