+++ livedata/src/test/kotlin/com/frybits/rx/preferences/livedata/Roshambo.kt
