+++ rx3/src/main/kotlin/com/frybits/rx/preferences/rx3/Optional.kt
